+++ web_ui/src/Chat.js
@@ -3,7 +3,7 @@ import { debounce } from "lodash";
 
 import { useEffect, useState, useContext, useCallback, useRef } from 'react';
 import { Card, Box, Paper, Toolbar, IconButton, Typography, TextField,
-    List, ListItem, Menu, MenuItem, Tooltip, Button
+    List, ListItem, Menu, MenuItem, Tooltip, Button, FormLabel, Popover
      } from '@mui/material';
 import { styled } from '@mui/system';
 import { ClassNames } from "@emotion/react";
@@ -102,11 +102,20 @@ const Chat = ({
 
     // AI Library state
     const [aiLibrary, setAiLibrary] = useState([]);
-    const [selectedAiLibraryNoteId, setSelectedAiLibraryNoteId] = useState(null);
-    const [selectedAiLibraryNoteName, setSelectedAiLibraryNoteName] = useState("");
-    const [selectedAiLibraryNoteContent, setSelectedAiLibraryNoteContent] = useState("");
-    const [showAiLibraryHelp, setShowAiLibraryHelp] = useState(false);
+    const [selectedAiLibraryNotes, setSelectedAiLibraryNotes] = useState({});
+    const [selectedAiLibraryNoteId, setSelectedAiLibraryNoteId] = useState("");
     const [aiLibraryOpen, setAiLibraryOpen] = useState(false);
+    const [showOnlyNotesInAiLibrary, setShowOnlyNotesInAiLibrary] = useState(true);
+    const [anchorAiLibraryHelpEl, setAnchorAiLibraryHelpEl] = useState(null);
+    const aiLibraryHelpPopoverOpen = Boolean(anchorAiLibraryHelpEl);
+
+    const handleAiLibraryHelpPopoverOpen = (event) => {
+        setAnchorAiLibraryHelpEl(event.currentTarget);
+      };
+      
+      const handleAiLibraryHelpPopoverClose = () => {
+        setAnchorAiLibraryHelpEl(null);
+      };
 
     const applyCustomSettings = () => {
         axios.get(`${serverUrl}/custom_settings/chat`).then(response => {
@@ -492,8 +501,13 @@ const Chat = ({
     const sendPrompt = async (prompt) => {
         // setup as much of the request as we can before calling appendMessage
         // as that will wait for any re-rendering and the id could change in that time
-        let knowledge = selectedAiLibraryNoteContent !== "" ? "Given the following knowledge: " + selectedAiLibraryNoteContent + "\n\n" : "";
-
+        let knowledge = "";
+        for (const [id, note] of Object.entries(selectedAiLibraryNotes)) {
+            knowledge += "KNOWLEDGE_NAME:" + note.metadata.name + "\nKNOWLEDGE_CONTENT:\n" + note.content.note + "\n\n";
+        }
+        if (knowledge !== "") {
+            knowledge = "Given the following knowledge:\n\n" + knowledge + "\n\nRespond to this prompt:\n\n";
+        }
         let requestData = {
             model_settings: myModelSettings,
             system_prompt: systemPrompt,
@@ -821,29 +835,6 @@ const Chat = ({
         setMarkdownRenderingOn(newSetting);
     };
 
-    const handleSelectAILibraryNote = (note) => {
-        setSelectedAiLibraryNoteId(note.id);
-        setSelectedAiLibraryNoteName(note.name);
-        console.log("Selected AI Library note: ", note);
-        if (note.id !== null) {
-            console.log("Chat loading AI Lubrary note", note.id);
-            axios.get(`${serverUrl}/docdb/notes/documents/${note.id}`, {
-                headers: {
-                    Authorization: 'Bearer ' + token
-                  }
-            }).then(response => {
-                console.log("Chat AI library note load Response", response);
-                response.data.access_token && setToken(response.data.access_token);
-                setSelectedAiLibraryNoteContent(response.data.content.note);
-            }).catch(error => {
-                console.log("Chat AI library note load error", error);
-                system.error(`Error loading Chat AI library note: ${error}`);
-            });
-        } else {
-            setSelectedAiLibraryNoteContent("");
-        }
-    }
-
     const handleToggleAILibraryOpen = () => {
         if (aiLibraryOpen) {
             setAiLibraryOpen(false);
@@ -853,8 +844,42 @@ const Chat = ({
         }
     }
 
+    const handleloadKnowledgeToAi = (event) => {
+        const noteStub = event.target.value;
+        if (noteStub && noteStub.id) {
+            console.log("Chat loading knowledge to AI", noteStub.id, noteStub.name);
+            axios.get(`${serverUrl}/docdb/notes/documents/${noteStub.id}`, {
+                headers: {
+                    Authorization: 'Bearer ' + token
+                  }
+            }).then(response => {
+                console.log("Chat AI library note load Response", response);
+                response.data.access_token && setToken(response.data.access_token);
+                const aiLibraryNote = response.data;
+                const updatedSelectedAiLibraryNotes = { ...selectedAiLibraryNotes, [aiLibraryNote.metadata.id]: aiLibraryNote };
+                setSelectedAiLibraryNotes(updatedSelectedAiLibraryNotes);
+                console.log("Selected aiLibrary notes", selectedAiLibraryNotes);
+                // reset the Select component
+                setSelectedAiLibraryNoteId("");
+            }).catch(error => {
+                console.log("Chat AI library note load error", error);
+                system.error(`Error loading Chat AI library note: ${error}`);
+            });
+        }
+    }
+
+    const handleUnloadKnowledgeFromAi = (id) => {
+        if (id && id in selectedAiLibraryNotes) {
+            console.log("Unloading knowledge from AI: ", id);
+            const updatedSelectedAiLibraryNotes = { ...selectedAiLibraryNotes };
+            delete updatedSelectedAiLibraryNotes[id];
+            setSelectedAiLibraryNotes(updatedSelectedAiLibraryNotes);
+            console.log("Selected aiLibrary notes", selectedAiLibraryNotes);
+        }
+    }
+
     const render = <Card id="chat-panel" sx={{display:"flex", flexDirection:"column", padding:"6px", margin:"6px", flex:1, minWidth: "400px", maxWidth: maxWidth ? maxWidth : "600px" }}>
-    <StyledToolbar className={ClassNames.toolbar} sx={{ gap: 1 } }>
+    <StyledToolbar className={ClassNames.toolbar} sx={{ gap: 1 }}>
         <CommentIcon/>
         <Typography sx={{mr:2}}>Chat</Typography>
         <Tooltip title={ id === "" ? "You are in a new chat" : "New chat"}>
@@ -991,7 +1016,7 @@ const Chat = ({
                 </ListItem>}
             </List>
         </Box>
-        <Box sx={{ minHeight: "128px" }}>
+        <Box sx={{ display: "flex", flexDirection: "column", minHeight: "128px" }}>
             <SecondaryToolbar className={ClassNames.toolbar} sx={{ gap: 1 }}>
                 <Typography sx={{mr:2}}>Prompt Editor</Typography>
                 <Tooltip title={ "Save prompt as template" }>
@@ -1002,10 +1027,10 @@ const Chat = ({
                         </IconButton>
                     </span>
                 </Tooltip>
-                <Tooltip title={ aiLibraryOpen ? "Hide AI Library" : `Show AI Library (Current: ${selectedAiLibraryNoteName})`}>
+                <Tooltip title={ aiLibraryOpen ? "Hide AI Library" : `Show AI Library (${Object.keys(selectedAiLibraryNotes).length} knowledge notes loaded)`}>
                     <IconButton edge="start" color="inherit" aria-label={ aiLibraryOpen ? "Hide AI Library" : "Show AI Library"}
                         onClick={handleToggleAILibraryOpen}>
-                        { selectedAiLibraryNoteName === "" ? <LocalLibraryOutlinedIcon/> : <LocalLibraryIcon/> }
+                        { Object.keys(selectedAiLibraryNotes).length === 0 ? <LocalLibraryOutlinedIcon/> : <LocalLibraryIcon/> }
                     </IconButton>
                 </Tooltip>
                 <Tooltip title={ "Ask again" }>
@@ -1046,7 +1071,7 @@ const Chat = ({
                 </Box>
             </SecondaryToolbar>
             <TextField 
-                sx={{ width: "100%", mt: "auto", overflow: "auto", maxHeight: "300px" }}
+                sx={{ width: "100%", mt: "auto", flex: 1, overflow: "auto", minHeight: "56px", maxHeight: "300px" }}
                     id="chat-prompt"
                     multiline 
                     variant="outlined" 
@@ -1075,28 +1100,69 @@ const Chat = ({
             </Paper>
             { aiLibraryOpen ? 
                 <Paper sx={{ margin: "2px 0px", padding: "2px 6px", display:"flex", gap: 1, backgroundColor: grey[100] }}>
-                    <FormControl sx={{ mt: 2, minWidth: 120, width: "100%" }}>
-                        <Box sx={{ display: "flex", direction: "row", width: "100%" }}>
-                            <InputLabel id="ai-library-helper-label">AI Library Note</InputLabel>
-                            <Select
-                            sx={{width: "100%"}}
-                            labelId="ai-library-select-helper-label"
-                            id="ai-library-select-helper"
-                            value={selectedAiLibraryNoteId}
-                            label="Selected AI Library Note"
-                            onChange={(event) => { handleSelectAILibraryNote({id: event.target.value, name: event.target.textContent}); }}
-                            >
-                                  <MenuItem value="None"><em>None</em></MenuItem>
-                                    {aiLibrary.map((note) => (
-                                        <MenuItem key={note.id} value={note.id}>{note.name}</MenuItem>
+                    <Box sx={{ mt: 2, display: "flex", flexDirection: "column", width: "100%" }}>
+                        <FormLabel>Loaded knowledge: { Object.keys(selectedAiLibraryNotes).length === 0 ? "None" : ""}</FormLabel>
+                        <List dense sx={{ width: "100%", overflow: "auto", maxHeight: "100px" }}>
+                            {Object.values(selectedAiLibraryNotes).map(note =>(
+                                <ListItem 
+                                    key={"loaded-ai-knowledge-" + note.metadata.id}
+                                    secondaryAction={
+                                        <Tooltip title={ "Unload knowledge from AI" }>
+                                            <IconButton edge="end" aria-label="Unload knowledge from AI" onClick={() => {handleUnloadKnowledgeFromAi(note.metadata.id)}}>
+                                                <CloseIcon />
+                                            </IconButton>
+                                        </Tooltip>
+                                      }
+                                >
+                                    {note.metadata.name}
+                                </ListItem>
+                            ))}
+                        </List>
+                        <FormControl sx={{ minWidth: 120, width: "100%" }}>
+                            <Box sx={{ display: "flex", direction: "row", width: "100%" }}>
+                                <InputLabel id="ai-library-helper-label">Select knowledge to add</InputLabel>
+                                <Select
+                                sx={{width: "100%"}}
+                                labelId="ai-library-select-helper-label"
+                                id="ai-library-select-helper"
+                                label="Select notes to add to AI knowledge"
+                                value={selectedAiLibraryNoteId}
+                                onChange={handleloadKnowledgeToAi}
+                                >
+                                    {(showOnlyNotesInAiLibrary ? aiLibrary.filter(noteStub => noteStub.properties.inAILibrary) : aiLibrary).map((noteStub) => (
+                                        <MenuItem key={'ai-library-item-' + noteStub.id} value={noteStub}>{noteStub.name}</MenuItem>
                                     ))}
-                            </Select>
-                            <IconButton onClick={() => { setShowAiLibraryHelp(x=>!x) }}>
+                                </Select>
+                                <Tooltip title={ showOnlyNotesInAiLibrary ? "Select box will show notes that are in the AI Library. Click to show all notes" : "Show only notes in AI library" }>
+                                    <IconButton onClick={() => { setShowOnlyNotesInAiLibrary(x=>!x) }}>
+                                        { showOnlyNotesInAiLibrary ? <LocalLibraryIcon/> : <LocalLibraryOutlinedIcon/> }
+                                    </IconButton>
+                                </Tooltip>
+                                <IconButton onClick={handleAiLibraryHelpPopoverOpen}>
                                 <HelpIcon />
-                            </IconButton>
-                        </Box>
-                        <FormHelperText>{ showAiLibraryHelp ? "Add knowledge the AI can use to answer questions by creating notes and clicking the book icon in the notes main toolbar to add them to the AI library, then select one from this list for the AI to use in answering your questions." : "" }</FormHelperText>
-                    </FormControl>
+                                </IconButton>
+                                <Popover
+                                id="ai-library-help-popover"
+                                open={aiLibraryHelpPopoverOpen}
+                                    anchorEl={anchorAiLibraryHelpEl}
+                                onClick={handleAiLibraryHelpPopoverClose}
+                                onClose={handleAiLibraryHelpPopoverClose}
+                                anchorOrigin={{
+                                    vertical: 'bottom',
+                                    horizontal: 'left',
+                                }}
+                                transformOrigin={{
+                                    vertical: 'top',
+                                    horizontal: 'left',
+                                }}
+                                >
+                                <FormHelperText sx={{ p: 2, width: "300px" }}>
+                                    Add knowledge the AI can use to answer questions by creating notes and clicking the book icon in the notes main toolbar to add them to the AI library, then select notes from this list for the AI to use in answering your questions. Click the X next to loaded knowledge notes to unload them. Loaded knowledge takes up space in the AI context window, so ensure it is concise.
+                                </FormHelperText>
+                                </Popover>
+                            </Box>
+                        </FormControl>
+                    </Box>
                 </Paper> : null
             }
         </Box>